requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
paddlepaddle==2.5.1
paddleocr==2.7.0.3
opencv-python-headless==4.8.1.78
pillow==10.1.0
numpy==1.24.3
